tests/python/requirements.txt
cx_Oracle
pytest
PyHamcrest
